requirements.txt
discord.py
pandas
ta
requests
pybit
pycryptodome
python-dotenv
